Add shared tests for parseAmount
