+++ composeApp/src/androidUnitTest/kotlin/de/lehrbaum/firefly/TestLocale.android.kt
@@ -1,0 +1,13 @@
+package de.lehrbaum.firefly
+
+import java.util.Locale
+
+actual fun withEnglishLocale(block: () -> Unit) {
+	val previousLocale = Locale.getDefault()
+	Locale.setDefault(Locale.US)
+	try {
+		block()
+	} finally {
+		Locale.setDefault(previousLocale)
+	}
+}
