+++ composeApp/src/commonTest/kotlin/de/lehrbaum/firefly/TestLocale.kt
@@ -1,0 +1,3 @@
+package de.lehrbaum.firefly
+
+expect fun withEnglishLocale(block: () -> Unit)
