+++ composeApp/src/iosArm64Test/kotlin/de/lehrbaum/firefly/TestLocale.ios.kt
@@ -1,0 +1,5 @@
+package de.lehrbaum.firefly
+
+actual fun withEnglishLocale(block: () -> Unit) {
+	block()
+}
